test_requirements.txt
-e .
colour-runner==0.0.4
django==1.6.2
dj_database_url==0.3.0
factory-boy==2.3.1
incuna-test-utils==2.1.0
psycopg2==2.5.2
rest-framework-generic-relations==0.1.0
